[ROCm][CI] Fix AITER MLA fp8 decode metadata regression test
The regression test in test_rocm_aiter_mla_decode_metadata.py was added in
#46997 together with dtype forwarding to get_mla_metadata_v1, when the decode
query dtype was bf16. A later bugfix (#47276) intentionally set the query dtype
to fp8 on the gfx950 DPA+FP8 KV decode path to fix a memory access fault, but
the regression test was not updated and started failing.

Align the test's expected query dtype with the current (correct) builder
behavior: fp8 KV-cache decode forwards an fp8 query dtype to
get_mla_metadata_v1. Update EXPECTED_Q_DTYPE to torch.float8_e4m3fn and the
accompanying comment.

+++ tests/kernels/attention/test_rocm_aiter_mla_decode_metadata.py
@@ -40,9 +40,10 @@ BATCH_SIZE = 128
 CONTEXT_LEN = 8192
 PAGE_SIZE = 1
 
-# Expected dtypes for this fold path: bf16 model dtype -> bf16 query; fp8
-# KV-cache -> fp8_e4m3 kv.
-EXPECTED_Q_DTYPE = torch.bfloat16
+# Expected dtypes for this fold path: fp8 KV-cache decode forwards an fp8 query
+# dtype to get_mla_metadata_v1 (see #47276, which fixes a memory access fault on
+# the gfx950 DPA+FP8 KV path); fp8 KV-cache -> fp8_e4m3 kv.
+EXPECTED_Q_DTYPE = torch.float8_e4m3fn
 EXPECTED_KV_DTYPE = torch.float8_e4m3fn
 
 # The split/reduce content tensors filled by get_mla_metadata_v1. work_meta_data
